requirements.txt
python-dotenv
sendgrid
tabulate